ci: streamline biome workflow
