+++ .github/workflows/biome.yml
@@ -8,9 +8,14 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
       - uses: oven-sh/setup-bun@v1
         with:
           bun-version: 1.2.14
       - run: bun install
-      - run: bun run lint
-      - run: bun x biome check .
+      - run: bun run lint:fix
+      - run: bun run format
+      - uses: lucasilverentand/auto-commit@v1
+        with:
+          commit-message: "chore: apply Biome fixes"
