+++ packages/core/src/base.ts
@@ -1,0 +1,3 @@
+import { z } from "zod";
+
+export const Base = z.object({ id: z.string().min(1) });
