+++ packages/dot-steward/src/plan.ts
@@ -1,0 +1,9 @@
+import type { Profile } from "./profile";
+
+export interface Plan {
+  profiles: Profile[];
+}
+
+export function plan(p: Plan): Plan {
+  return p;
+}
