+++ plugins/apt/src/plugin.ts
@@ -1,0 +1,9 @@
+import { Plugin } from "@dot-steward/core";
+import { AptItem, type Apt } from "./item";
+
+export class AptPlugin extends Plugin<Apt> {
+  name = "apt";
+  schema = AptItem;
+}
+
+export const plugin = new AptPlugin();
